+++ src/Vortice.Direct2D1/WIC/IWICBitmapLock.cs
@@ -1,6 +1,7 @@
 // Copyright (c) Amer Koleci and contributors.
 // Distributed under the MIT license. See the LICENSE file in the project root for more information.
 
+using System;
 using Vortice.Mathematics;
 
 namespace Vortice.WIC
@@ -16,12 +17,28 @@ namespace Vortice.WIC
             }
         }
 
-        public byte[] GetDataPointer()
+        /// <summary>
+        /// Gets a <see cref="DataRectangle"/> to the data.
+        /// </summary>
+        public DataRectangle Data
+        {
+            get
+            {
+                var pointer = GetDataPointer(out var size);
+                return new DataRectangle(pointer, Stride);
+            }
+        }
+
+        public unsafe Span<byte> GetDataPointer() 
         {
             var pointer = GetDataPointer(out var size);
-            var data = new byte[size];
-            Interop.Read(pointer, data);
-            return data;
+            return new Span<byte>(pointer.ToPointer(), size);
+        }
+
+        public unsafe Span<T> GetDataPointer<T>() where T : unmanaged
+        {
+            var pointer = GetDataPointer(out var size);
+            return new Span<T>(pointer.ToPointer(), size / sizeof(T));
         }
     }
 }