+++ CHANGELOG.md
@@ -7,6 +7,7 @@ Release:     1.7.0 (May 2020)
 -----------------------------------------------
 KEY CHANGES:
 - [ENH] DXGI: Expose types from windows.ui.xaml.media.dxinterop.h
+- [FIX] WIC: Crash with IWICBitmapLock.GetDataPointer, correctly handle different types
 
 -----------------------------------------------
 Release:     1.6.0 (April 2020)